### 2023-10-16 a903d15
removed bug for convolution with small psf sizes
