+++ shopt.bib
@@ -1,0 +1,9 @@
+@misc{berman2023shoptjl,
+      title={ShOpt.jl: A Julia Package for Empirical Point Spread Function Characterization of JWST NIRCam Data}, 
+      author={Edward Berman and Jacqueline McCleary},
+      year={2023},
+      eprint={2310.00071},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.IM}
+}
+
