### 2023-12-12 150e3b9
changed initialization
