### 2023-12-13 #6144 @editorialbot
[REVIEW]: ShOpt.jl: A Julia Package for Empirical Point Spread Function Characterization of JWST NIRCam Data
**Submitting author:** @EdwardBerman (<a href="http://orcid.org/0000-0002-8494-3123">Edward Berman</a>)
**Repository:** https://github.com/EdwardBerman/shopt
**Branch with paper.md** (empty if default branch):
**Version:** v3.0.1
