### 2023-12-17 EdwardBerman/shopt#7 @kevinmicha
[JOSS] Paper comments
Hi,

Your [JOSS paper](https://github.com/openjournals/joss-papers/blob/joss.06144/joss.06144/10.21105.joss.06144.pdf) is pretty well written and clear! I have some minor suggestions, including fixing a few typos:

- I would remove the **Introduction** title, leaving simply **Summary**;
- Typo: 'unprecendented' (line 35);
- Typo: ‘biasses’ (line 67);
- Typo: ‘emblamatic’ (line 76);
- For the State of the Field section: I wonder how `ShOpt.jl` would compare with `PSFr` (https://github.com/sibirrer/psfr), which is quite recent and has been used for JWST data;
- **Beichman et al., 2012a** and **Beichman et al., 2012b** are exactly the same paper.
