### 2023-12-17 EdwardBerman/shopt#8 @kevinmicha
[JOSS] Reproducibility: Figure 2 
Hi,

I am trying to reproduce Figure 2 of the [paper](https://github.com/openjournals/joss-papers/blob/joss.06144/joss.06144/10.21105.joss.06144.pdf), and while it appears visually appealing, it would be helpful to have some additional clarifications:

1. It would be beneficial if you could specify in the repository (e.g. `README`) or on the GitHub site the precise JWST data file you downloaded and whether any modifications were made to the default `SExtractor` configuration files.
2. Is it possible for users to generate Figure 2 by simply providing the exact catalog and using `TutorialNotebook.ipynb`, or would additional code be required?

Thanks a lot for your time!
