### 2023-12-17 41c519f
added idx to SNR_WIN
