### 2023-12-17 #6144 @kevinmicha
Hi @EdwardBerman, I have been reviewing the `ShOpt.jl` repository (and paper), and I must say this is a solid contribution to the PSF modelling field!

I have opened two issues in your repository so far; please let me know when they are fixed so I can continue to check the boxes:

[[JOSS] Reproducibility: Figure 2](https://github.com/EdwardBerman/shopt/issues/8)
[[JOSS] Paper comments](https://github.com/EdwardBerman/shopt/issues/7)
