+++ paper.bib
@@ -437,3 +437,81 @@ abstract = {The Astropy Project supports and fosters the development of open-sou
       archivePrefix={arXiv},
       primaryClass={astro-ph.GA}
 }
+
+@ARTICLE{2023ApJ942L27S,
+       author = {{Santini}, P. and {Fontana}, A. and {Castellano}, M. and {Leethochawalit}, N. and {Trenti}, M. and {Treu}, T. and {Belfiori}, D. and {Birrer}, S. and {Bonchi}, A. and {Merlin}, E. and {Mason}, C. and {Morishita}, T. and {Nonino}, M. and {Paris}, D. and {Polenta}, G. and {Rosati}, P. and {Yang}, L. and {Boyett}, K. and {Bradac}, M. and {Calabr{\`o}}, A. and {Dressler}, A. and {Glazebrook}, K. and {Marchesini}, D. and {Mascia}, S. and {Nanayakkara}, T. and {Pentericci}, L. and {Roberts-Borsani}, G. and {Scarlata}, C. and {Vulcani}, B. and {Wang}, Xin},
+        title = "{Early Results from GLASS-JWST. XI. Stellar Masses and Mass-to-light Ratio of z > 7 Galaxies}",
+      journal = {\apjl},
+     keywords = {High-redshift galaxies, Stellar masses, Mass-to-light ratio, 734, 1614, 1011, Astrophysics - Astrophysics of Galaxies},
+         year = 2023,
+        month = jan,
+       volume = {942},
+       number = {2},
+          eid = {L27},
+        pages = {L27},
+          doi = {10.3847/2041-8213/ac9586},
+archivePrefix = {arXiv},
+       eprint = {2207.11379},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023ApJ...942L..27S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{2022ApJ938L14M,
+       author = {{Merlin}, Emiliano and {Bonchi}, Andrea and {Paris}, Diego and {Belfiori}, Davide and {Fontana}, Adriano and {Castellano}, Marco and {Nonino}, Mario and {Polenta}, Gianluca and {Santini}, Paola and {Yang}, Lilan and {Glazebrook}, Karl and {Treu}, Tommaso and {Roberts-Borsani}, Guido and {Trenti}, Michele and {Birrer}, Simon and {Brammer}, Gabriel and {Grillo}, Claudio and {Calabr{\`o}}, Antonello and {Marchesini}, Danilo and {Mason}, Charlotte and {Mercurio}, Amata and {Morishita}, Takahiro and {Strait}, Victoria and {Boyett}, Kristan and {Leethochawalit}, Nicha and {Nanayakkara}, Themiya and {Vulcani}, Benedetta and {Bradac}, Marusa and {Wang}, Xin},
+        title = "{Early Results from GLASS-JWST. II. NIRCam Extragalactic Imaging and Photometric Catalog}",
+      journal = {\apjl},
+     keywords = {Surveys, Photometry, Galaxies, Astronomy image processing, 1671, 1234, 573, 2306, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2022,
+        month = oct,
+       volume = {938},
+       number = {2},
+          eid = {L14},
+        pages = {L14},
+          doi = {10.3847/2041-8213/ac8f93},
+archivePrefix = {arXiv},
+       eprint = {2207.11701},
+ primaryClass = {astro-ph.GA},
[…]
