+++ paper.md
@@ -50,7 +50,7 @@ Empirical PSF characterization tools like PSFEx [@2011ASPC] and PIFF [@Jarvis_20
 `ShOpt` bridges the speed of `PSFex` with the features of `PIFF` using fewer configurable hyperparameters. `ShOpt` employs a myriad of techniques to optimize the speed of the program. First and foremost, `ShOpt` is equipped with support for multithreading. Polynomial interpolation is used for handling PSF variations across the field of view. The polynomials given to each basis element of the PSF are independent of one another and therefore can be distributed to different CPU threads to be run in parallel. `ShOpt` also introduces new methods for fitting both analytic profiles and pixel based profiles. If an analytic profile is used to model the PSF, then there are $3$ basis elements parameterizing the model. We choose 2D elliptical Gaussians for these analytic profiles because they are cheap to compute. Moreover, we use the Limited Memory Broyden–Fletcher–Goldfarb–Shanno algorithm (LBFGS) to find these paramters. This is faster but more memory intensive than Conjugate Gradient, the algorithm used in `PIFF`. Moreover, the $3$ basis elements are constrained to the manifold $B_2(r) \times \mathbb{R}_+$. We constructed a function that maps any point in $\mathbb{R}^3$ into $B_2(r) \times \mathbb{R}_+$. The LBFGS algorithm uses successive iterations to converge to a solution for the $3$ basis elements, and so we use this function to ensure that our update steps in LBFGS do not leave the constraint. For the pixel basis, both `PIFF` and `PSFex` approximate the PSF by minimizing the reduced $\chi^2$ between a grid of pixels and a star vignette. PCA can quickly achieve the same purpose of approximating the input vignette without overfitting to background noise. We also provide an autoencoder mode, which uses deep learning to reconstruct the image. The weights and biases are not reset between stars, so the knowledge of how to reconstruct one star is transfered to the next. This in turn leads to less training iterations. Finally, `ShOpt` is written in Julia. Julia uses a just in time compiler which makes it faster than intepreted languages such as Python and has shown to be a good choice for performance critical code [@Stanitzki_2021]. 
 
 # State of the Field
-The JWST captures images at high resolution and at wavelengths of light that have been previously unexplored [@Gardner_2006]. With these images we are seeing farther into the early universe than we ever have before. The difficulties of producing good PSF models for the JWST are emblametic of a larger problem: Our data sets are getting bigger and existing software was not built to scale. That is to say, the advancements in software are falling behind the advances in instrumentation. Not only does `ShOpt` produces PSF models for JWST NIRCam images, it also sets the precedent for designing software that scales.  
+The JWST captures images at high resolution and at wavelengths of light that have been previously unexplored [@Gardner_2006]. With these images we are seeing farther into the early universe than we ever have before. The difficulties of producing good PSF models for the JWST are emblemetic of a larger problem: Our data sets are getting bigger and existing software was not built to scale. That is to say, the advancements in software are falling behind the advances in instrumentation. Not only does `ShOpt` produces PSF models for JWST NIRCam images, it also sets the precedent for designing software that scales.  
 
[…]
