### 2023-12-23 EdwardBerman/shopt#7 @EdwardBerman
Thank you for your comments! I just pushed a paper with these typos and suggestions incorporated. I've seen your other issue as well, I'm planning on addressing that over my holiday break.
