### 2023-12-23 EdwardBerman/shopt#7 @kevinmicha
Thank you @EdwardBerman! I'm happy with most of the changes, I'm just checking that you've replaced _emblamatic_ with _emblemetic_, but it's really minor.

Once you've changed that last aspect, this issue can be closed (feel free to do it).
