### 2024-01-04 EdwardBerman/shopt#8 @EdwardBerman
Hello, thank you for these considerations!

I've added a test catalog and a `SExtractor` config file, as well as some default `PIFF` and `PSFex` configs that I've been using for benchmarking. Figure 2 in the paper won't directly be made by `TutorialNotebook.ipynb`, only the renderings for the `ShOpt` PSFs at specified coordinates. However, I've gone ahead and updated the notebook with this new test catalog to make it easier to test functionality. Assuming you've gone through the instillation procedure and configured PyCall on your system, I have tested that the tutorial notebooks runs smoothly. Note that you will still have to modify the file paths to that of your system and such. In evaluating `ShOpt` I've used all of the wavelengths with the 30mas pixel scale at the provided link and I've added that as a note to the `README`.
