### 2024-01-05 EdwardBerman/shopt#8 @kevinmicha
Hi @EdwardBerman, thank you for adding the requested files, clarifying some details and updating the notebook.

I was able to successfully run `ShOpt` for both `PCA` and `autoencoder` modes. The chi-squared pixel grid fit mode fails, but it is listed in the `README` Known Issues and Enhancements section, so this is fine to me.

I think I am ready to check all the items on the review list - but I just checked that _emblematic_ is still spelled incorrectly in the paper. When you fix that, would you close this issue and type `@editorialbot generate pdf` [here](https://github.com/openjournals/joss-reviews/issues/6144) so the PDF is regenerated? Thanks a lot!
