### 2024-01-05 #6144 @kevinmicha
Hi,

@EdwardBerman et al. have solved the issues I opened via the Dec 23, 2023 and Jan 4, 2024 [commits](https://github.com/EdwardBerman/shopt/commits/main/).

@ivastar I have ticked all the boxes and am happy to recommend `ShOpt` for publication in JOSS. Well done @EdwardBerman et al.!
