### 2024-01-13 #6144 @jpierel14
@ivastar Sorry can you point me to where the most recent version of the paper is? I read the version linked above, but then noticed in [here](https://github.com/EdwardBerman/shopt/issues/7) that the paper had been updated. I see in the repo, the pdf has not been updated but the markdown version of the paper has, and does not match the pdf linked above by the bot. I have read the paper though and am ready to give comments, just want to make sure I'm on the most recent version!
