### 2024-01-15 EdwardBerman/shopt#9 @aymgal
[JOSS] Improving the documentation
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6144) process.

I have been following the installation instructions and running of the program on the test catalog. Here are some comments regarding the documentation.

- The part on how to compile Python libraries for use in Julia is not straightforward to follow. In particular, it is not clear that there are three, even four, ways to achieve the same thing. I advise to re-formulate this part, stating ahead that there are these X different ways (e.g., as a list), and guide the user to the right section, instead of having all the commands in a row.
- It is important that the command `pyimport("astropy")` should be run _after_ re-loading the Julia REPL, so it would better to have this explicitely mentioned.
- Is the `TutorialNotebook.ipynb` made just for copy-pasting lines of codes into the Julia REPL, or can this be executed using Julia directly? This is not clear to me. Probably because this is the first time I run Julia code: it means that the documentation can be be improved.
- `using Base.Threads` is missing in `TutorialNotebook.ipynb`.
