### 2024-01-15 EdwardBerman/shopt#10 @aymgal
[JOSS] Missing references
In the current state of the JOSS paper, I suggest to mention [`STARRED`](https://joss.theoj.org/papers/10.21105/joss.05340) as another efficient PSF reconstruction technique, as it has been shown to outperform other programs like PSFex (mentioned in the paper) and PSFr (which would need to be mentioned as well), and uses high the just-in-time compiled and auto-diff library JAX.
