### 2024-01-15 EdwardBerman/shopt#11 @aymgal
[JOSS] Default output directory naming on macos
The default naming of the output directory follows something like `2024-01-15T11:20:40.848`, but `:` characters are forbidden in macos path names (and are replaced by the `/` when displayed in a file explorer). Hence I would suggest not to use these characters but `-` or `_` instead.
