### 2024-01-15 #6144 @aymgal
## Review ("Functionality" only)

### Conflict of interest

- [x] I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

### Code of Conduct

- [x] I confirm that I read and will adhere to the JOSS code of conduct.

### Comments

#### The setup used for these tests:

- Laptop: Apple MacBook Pro M2, macos Ventura 13.2.1
- Python version: 3.10.13
- Julia version: 1.10.0

#### Issues opened

- [Improvements to documentation](https://github.com/EdwardBerman/shopt/issues/9)
- [Missing references](https://github.com/EdwardBerman/shopt/issues/10)
- [Output directory name](https://github.com/EdwardBerman/shopt/issues/11)

#### Questions to be answered

- Increasing `JULIA_NUM_THREADS`` does not reduce runtime for the example data set. Is that expected?

- I would like to see an easy way to retrieve a FITS file of the model of the PSF, centered on a cutout with a chosen number of pixels. There might already be a way to do this, but then it should be mentioned more clearly. This is important as many modeling programs make use of a pixelated kernel, with odd or even cutout size, centered on a pixel or in between pixels, etc.

@EdwardBerman Once the above points have been addressed, I am happy to tick the remaining boxes for the review :)

#### Extra comments:

- I am happy to report that the programs, following the documentation, runs without errors for the example data set. I tested it both with and without `export JULIA_NUM_THREADS=X`` (from 1 to 8 threads).

- The very verbose output of the program is really useful. Given that 136 stars are (initially) used when running the program with `test_catalog.fits`, my feeling is that indeed, ShOpt is very fast!
