### 2024-01-15 EdwardBerman/shopt#11 @EdwardBerman
This has been fixed with this  commit: [fbe320ab3cc4ddf4ae19daa8fa8a35819952880d](https://github.com/EdwardBerman/shopt/commit/fbe320ab3cc4ddf4ae19daa8fa8a35819952880d)
