+++ paper.bib
@@ -514,4 +514,18 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@article{michalewicz2023starred,
+  title={STARRED: a two-channel deconvolution method with Starlet regularization},
+  author={Michalewicz, Kevin and Millon, Martin and Dux, Fr{\'e}d{\'e}ric and Courbin, Fr{\'e}d{\'e}ric},
+  journal={Journal of Open Source Software},
+  volume={8},
+  number={85},
+  pages={5340},
+  year={2023}
+}
+@article{schoenholz2019jax,
+  title={Jax md: End-to-end differentiable, hardware accelerated, molecular dynamics in pure python},
+  author={Schoenholz, Samuel S and Cubuk, Ekin D},
+  year={2019}
+}
 
