### 2024-01-15 EdwardBerman/shopt#9 @EdwardBerman
I am continuing to work on the first two bullets. For the latter two:

- The TutorialNotebook is made to give the user a sense of order in which things should be done. Things can be executed within the notebook or copy and pasted into Julia desired. I clarify this with this commit: [96e97be](https://github.com/EdwardBerman/shopt/commit/96e97be360d257f4fe410d53508424dcf9b75d98)
- "using Base.Threads is missing in TutorialNotebook.ipynb." This is now fixed with commits [bb278d6](https://github.com/EdwardBerman/shopt/commit/bb278d6c3709323724689bdcf226434fa9256570)  [85e67c18a4cf420c9c721a24b985613516bde6a7](https://github.com/EdwardBerman/shopt/commit/85e67c18a4cf420c9c721a24b985613516bde6a7)
