### 2024-01-15 EdwardBerman/shopt#10 @aymgal
Thanks @EdwardBerman !

However, STARRED has also been tested on JWST data, see this [example notebook](https://gitlab.com/cosmograil/starred/-/blob/main/notebooks/more_examples/7_JWST-PSF_reconstruction.ipynb?ref_type=heads). So I would simply remove the mention about the HST data validation, since the PSF modeling assumptions behind the STARRED method are not specific to a given telescope.

In addition, there seems to be a typo "...PSF models with competitive PSFr and PSFex..."
