### 2024-01-15 EdwardBerman/shopt#10 @EdwardBerman
Ah, you are correct! Thank you @aymgal . I have fixed this with the following commit: [c388375](https://github.com/EdwardBerman/shopt/commit/c388375db62be6592a3d869ca16661b583b1d6a0)
