### 2024-01-15 #6144 @EdwardBerman
Extra comments and Questions for @aymgal :

- "Increasing `JULIA_NUM_THREADS`` does not reduce runtime for the example data set. Is that expected?"

The main speed up you should expect is during the interpolation across the field of view when you see "Transforming (x,y) -> (u,v) | Interpolation Pixel by Pixel Across the Field of View." For around 100 stars this should already be pretty fast, so it may be easy to miss.

- "I would like to see an easy way to retrieve a FITS file of the model of the PSF, centered on a cutout with a chosen number of pixels. There might already be a way to do this, but then it should be mentioned more clearly. This is important as many modeling programs make use of a pixelated kernel, with odd or even cutout size, centered on a pixel or in between pixels, etc."

This isn't a current way to export this information as a fits file, only to retrieve the PSF model at a give position from a summary fits file as a matrix in Julia. Additionally, there isn't currently a method to adjust the size of the output kernel, but this could be implemented in the future.

- "The very verbose output of the program is really useful."

This was encouraged by my co-author @mcclearyj , so thank you for this idea!
