### 2024-01-15 EdwardBerman/shopt#9 @EdwardBerman
The first two bullets have been addressed in commits [ba6d571](https://github.com/EdwardBerman/shopt/commit/ba6d571be5521b3e2ccbb43578e3b3c58e422264) through [6b4e764](https://github.com/EdwardBerman/shopt/commit/6b4e764bdc93e82f01b6bee88db401099556835f)!
