### 2024-01-16 EdwardBerman/shopt#10 @aymgal
Thanks for the update, I close this issue!
