### 2024-01-16 EdwardBerman/shopt#9 @aymgal
Great, I'm happy with these changes, I close this issue :)
