### 2024-01-16 #6144 @aymgal
> This isn't a current way to export this information as a fits file, only to retrieve the PSF model at a give position from a summary fits file as a matrix in Julia. Additionally, there isn't currently a method to adjust the size of the output kernel, but this could be implemented in the future.

Ok, thanks for your thoughts! I think it would be a very valuable feature. Hence if it is not too much of a burden to add this functionality (maybe "just" retrieving the a fits file of the PSF kernel, without choosing the kernel size) I would strongly suggest to do this. And mention this feature in the doc and paper as well.

But since I am not familiar enough with Julia and its interaction with FITS files, I obviously won't make make this as a blocking factor for acceptance :) So @EdwardBerman let me know what you think, and I will tick the last box!

All my other points above were adequately addressed by the author.
