### 2024-01-16 #6144 @jpierel14
@EdwardBerman I've left comments in this [Issue](https://github.com/EdwardBerman/shopt/issues/12), good work!
