### 2024-01-18 EdwardBerman/shopt#12 @EdwardBerman
I've addressed all of these comments in my latest rendition of the draft, and I have rerendered the PDF in the JOSS issue. To illustrate the non Gaussian nature of the PSF, I added a comparison between an idealized WebbPSF forward model and a Gaussian approximation. Thank you for your insights!
