### 2024-01-18 #6144 @EdwardBerman
@jpierel14 I have addressed the comments in the issue you left on the ShOpt repository, let me know if you have any further questions!
