### 2024-01-20 b5bc64e
diagnostic plot
