### 2024-01-25 #6144 @ivastar
@EdwardBerman I noticed that you have a separate submission to ApJ: https://arxiv.org/pdf/2401.11625.pdf

We actually have a special process for joint JOSS + ApJ submissions: https://joss.readthedocs.io/en/latest/editing.html#handling-of-papers-published-together-with-aas-publishing

I would highly recommend that we move to that process so the two papers are more explicitly liked. You are already citing the JOSS submission in the ApJ manuscript. This process will add the ApJ DOI to the JOSS paper.

Let me know if you have any objections.
