### 2024-01-25 #6144 @EdwardBerman
Thanks @ivastar , can I ask what that would entail? I thought what I had in the markdown with the AAS collaboration had already indicated that? If I wanted to move to that process, would I have to go through the full review process again?
