### 2024-01-25 #6144 @ivastar
@EdwardBerman no you do not have to do the review process again. The submission has been reviewed. It is possibly an oversight on our end that we did not note this as a AAS collaboration. I think all this entails is that we will pause the publication process and wait for the acceptance of the ApJ submission before we finalize the JOSS publication.
