### 2024-01-25 #6144 @EdwardBerman
I'd like some clarification before going through with this process. You have said that this will add the ApJ DOI to the JOSS paper, is this just a matter of adding the DOI to the watermark? If that's all it is, I may be inclined just to proceed with publication. Let me know if I'm missing anything! Also, note that we submitted to AJ, not ApJ, if that makes any difference.
