### 2024-01-26 #6144 @ivastar
> I'd like some clarification before going through with this process. You have said that this will add the ApJ DOI to the JOSS paper, is this just a matter of adding the DOI to the watermark? If that's all it is, I would prefer just to proceed with publication. Let me know if I'm missing anything! Also, note that we submitted to AJ, not ApJ, if that makes any difference.

It adds the AJ DOI to the text in the left column, yes, and JOSS gets a small financial contribution form AAS. Here is a description of the collaboration: https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
