### 2024-01-27 #6144 @EdwardBerman
@ivastar thank you for the clarification! I think that at this stage I would like to forgo this process and have the JOSS paper published, is that still possible? Otherwise we can wait for the AJ acceptance and proceed from there.
