+++ configdir/shopt.yml
@@ -26,7 +26,7 @@ plots: #What plots do you want?
     parametersHistogram: false
     parametersScatterplot: false
   cairomakiePlots:
-    streamplots: false
+    streamplots: true
   pythonPlots: false
 
 polynomialDegree: 1 #degree of polynomial for interpolation
