### 2024-01-31 #6144 @dfm
@EdwardBerman — @ivastar has asked me as the Associate Editor-in-Chief of this track to take a closer look at this discussion, so I wanted to weigh in. Given that the submission to the AAS Journals seems to also be focused on the ShOpt.jl software itself, I think that publishing this JOSS submission without connecting it to the AAS paper would be a case of multiple publication of the same result. The idea with the AAS collaboration is that JOSS can provide a code review in a way that the AAS Journals are not equipped to, even when both submissions are software focused. With this in mind, I'd like to second @ivastar's request that this submission be co-published under the AAS collaboration.

Our next steps will be:

1. @ivastar will lead us through the [post-review checklist](https://joss.readthedocs.io/en/latest/editorial_bot.html#post-review-checklist),
2. @EdwardBerman will share the JOSS DOI (it will be 10.21105/joss.06144 for this paper) with the AAS editor
3. We will pause the submission until the AAS submission is accepted
4. At which point, we'll finish the acceptance process, which only takes a few minutes!

Thanks all!
