+++ paper.md
@@ -39,7 +39,7 @@ Shear Optimization with `ShOpt.jl` introduces modern techniques, tailored to Jam
 ![ShOpt Workflow](flowchart.png)
 
 # Statement of need
-Empirical PSF characterization tools like Point Spread Function Extractor \citep[PSFEx, ][]{@2011ASPC} and Point Spread Functions in the Full Field of View \citep[PIFF, ][]{@Jarvis_2020} are widely popular in astrophysics. However, the quality of PIFF and PSFEx models tends to be quite sensitive to the parameter values used to run the software, with optimization sometimes relying on brute-force guess-and-check runs. PIFF is also notably inefficient for large, well-sampled images, taking hours in the worst cases. The JWST's Near Infrared Camera (NIRCam) offers vast scientific opportunities (e.g. [@casey2023cosmosweb]); at the same time, this unprecedented data brings new challenges for PSF modeling:
+Empirical PSF characterization tools like Point Spread Function Extractor (PSFEx [@2011ASPC]) and Point Spread Functions in the Full Field of View (PIFF [@Jarvis_2020]) are widely popular in astrophysics. However, the quality of PIFF and PSFEx models tends to be quite sensitive to the parameter values used to run the software, with optimization sometimes relying on brute-force guess-and-check runs. PIFF is also notably inefficient for large, well-sampled images, taking hours in the worst cases. The JWST's Near Infrared Camera (NIRCam) offers vast scientific opportunities (e.g. [@casey2023cosmosweb]); at the same time, this unprecedented data brings new challenges for PSF modeling:
 
 (1) Analytic functions like Gaussians are incomplete descriptions of the NIRCam PSF, as evident from Figures 2 and 3.  This calls for well-thought-out, non-parametric modeling and diagnostic tools that can capture the full dynamic range of the NIRCam PSF. `ShOpt` provides these models and diagnostics out of the box.
 
@@ -49,7 +49,7 @@ Empirical PSF characterization tools like Point Spread Function Extractor \citep
 
 ![The plot on the left shows an idealized forward model of the NIRCam F150W PSF made using WebbPSF. The middle shows a Gaussian approximation to the PSF. The right shows the residual between the WebbPSF model and the Gaussian approximation.](webbpsf_vs_gaussian.png)
 
[…]
