### 2024-02-25 #6144 @EdwardBerman
Hi @ivastar , I am working through these changes and I wanted to ask you about the first one regarding the \citep{} command. I am a little confused since I am preparing this with the markdown template which is then converted to latex, and running your command seemed to throw off the pandoc conversion. Could you add some clarification?
