### 2024-02-25 #6144 @EdwardBerman
I'm also a little unsure how I might switch the order of what's in the references, since that is done automatically upon reading my paper.bib. Besides that, thank you for these comments! They make the paper much better.
