+++ paper.bib
@@ -124,15 +124,23 @@ abstract = {We present the theoretical and analytical bases of optimal technique
 	journal = {Monthly Notices of the Royal Astronomical Society}
 }
 
-@article{casey2023cosmosweb,
-  title={COSMOS-Web: an overview of the JWST Cosmic Origins Survey},
-  author={Casey, Caitlin M and Kartaltepe, Jeyhan S and Drakos, Nicole E and Franco, Maximilien and Harish, Santosh and Paquereau, Louise and Ilbert, Olivier and Rose, Caitlin and Cox, Isabella G and Nightingale, James W and others},
-  journal={The Astrophysical Journal},
-  volume={954},
-  number={1},
-  pages={31},
-  year={2023},
-  publisher={IOP Publishing}
+@ARTICLE{casey2023cosmosweb,
+       author = {{Casey}, Caitlin M. and {Kartaltepe}, Jeyhan S. and {Drakos}, Nicole E. and {Franco}, Maximilien and {Harish}, Santosh and {Paquereau}, Louise and {Ilbert}, Olivier and {Rose}, Caitlin and {Cox}, Isabella G. and {Nightingale}, James W. and {Robertson}, Brant E. and {Silverman}, John D. and {Koekemoer}, Anton M. and {Massey}, Richard and {McCracken}, Henry Joy and {Rhodes}, Jason and {Akins}, Hollis B. and {Allen}, Natalie and {Amvrosiadis}, Aristeidis and {Arango-Toro}, Rafael C. and {Bagley}, Micaela B. and {Bongiorno}, Angela and {Capak}, Peter L. and {Champagne}, Jaclyn B. and {Chartab}, Nima and {Ch{\'a}vez Ortiz}, {\'O}scar A. and {Chworowsky}, Katherine and {Cooke}, Kevin C. and {Cooper}, Olivia R. and {Darvish}, Behnam and {Ding}, Xuheng and {Faisst}, Andreas L. and {Finkelstein}, Steven L. and {Fujimoto}, Seiji and {Gentile}, Fabrizio and {Gillman}, Steven and {Gould}, Katriona M.~L. and {Gozaliasl}, Ghassem and {Hayward}, Christopher C. and {He}, Qiuhan and {Hemmati}, Shoubaneh and {Hirschmann}, Michaela and {Jahnke}, Knud and {Jin}, Shuowen and {Khostovan}, Ali Ahmad and {Kokorev}, Vasily and {Lambrides}, Erini and {Laigle}, Clotilde and {Larson}, Rebecca L. and {Leung}, Gene C.~K. and {Liu}, Daizhong and {Liaudat}, Tobias and {Long}, Arianna S. and {Magdis}, Georgios and {Mahler}, Guillaume and {Mainieri}, Vincenzo and {Manning}, Sinclaire M. and {Maraston}, Claudia and {Martin}, Crystal L. and {McCleary}, Jacqueline E. and {McKinney}, Jed and {McPartland}, Conor J.~R. and {Mobasher}, Bahram and {Pattnaik}, Rohan and {Renzini}, Alvio and {Rich}, R. Michael and {Sanders}, David B. and {Sattari}, Zahra and {Scognamiglio}, Diana and {Scoville}, Nick and {Sheth}, Kartik and {Shuntov}, Marko and {Sparre}, Martin and {Suzuki}, Tomoko L. and {Talia}, Margherita and {Toft}, Sune and {Trakhtenbrot}, Benny and {Urry}, C. Megan and {Valentino}, Francesco and {Vanderhoof}, Brittany N. and {Vardoulaki}, Eleni and {Weaver}, John R. and {Whitaker}, Katherine E. and {Wilkins}, Stephen M. a
[…]
