### 2024-03-18 #6144 @EdwardBerman
@ivastar Thank you for this link, I will fix the citations using this and I will investigate the DOI issue. The AAS submission is still in review. They are being a little slow, but I will message here when there are updates. Thank you for your patience and for these editorial comments!
