+++ paper.bib
@@ -1,8 +1,16 @@
 @article{mccleary2023lensing,
-  title={Lensing in the Blue II: Estimating the Sensitivity of Stratospheric Balloons to Weak Gravitational Lensing},
-  author={McCleary, Jacqueline E and Everett, Spencer W and Shaaban, Mohamed M and Gill, Ajay S and Vassilakis, Georgios N and Huff, Eric M and Massey, Richard J and Benton, Steven J and Brown, Anthony M and Clark, Paul and others},
-  journal={arXiv preprint arXiv:2307.03295},
-  year={2023}
+doi = {10.3847/1538-3881/ace7ca},
+url = {https://dx.doi.org/10.3847/1538-3881/ace7ca},
+year = {2023},
+month = {aug},
+publisher = {The American Astronomical Society},
+volume = {166},
+number = {3},
+pages = {134},
+author = {Jacqueline E. McCleary and Spencer W. Everett and Mohamed M. Shaaban and Ajay S. Gill and Georgios N. Vassilakis and Eric M. Huff and Richard J. Massey and Steven J. Benton and Anthony M. Brown and Paul Clark and Bradley Holder and Aurelien A. Fraisse and Mathilde Jauzac and William C. Jones and David Lagattuta and Jason S.-Y. Leung and Lun Li and Thuy Vy T. Luu and Johanna M. Nagy and C. Barth Netterfield and Emaad Paracha and Susan F. Redmond and Jason D. Rhodes and Jürgen Schmoll and Ellen Sirks and Sut Ieng Tam},
+title = {Lensing in the Blue. II. Estimating the Sensitivity of Stratospheric Balloons to Weak Gravitational Lensing},
+journal = {The Astronomical Journal},
+abstract = {The Superpressure Balloon-borne Imaging Telescope (SuperBIT) is a diffraction-limited, wide-field, 0.5 m, near-infrared to near-ultraviolet observatory designed to exploit the stratosphere’s space-like conditions. SuperBIT’s 2023 science flight will deliver deep, blue imaging of galaxy clusters for gravitational lensing analysis. In preparation, we have developed a weak-lensing measurement pipeline with modern algorithms for PSF characterization, shape measurement, and shear calibration. We validate our pipeline and forecast SuperBIT survey properties with simulated galaxy cluster observations in SuperBIT’s near-UV and blue bandpasses. We predict imaging depth, galaxy number (source) density, and redshift distribution for observations in SuperBIT’s three bluest filters; the effect of lensing sample selections is also considered. We find that, in three hours of on-sky integration, SuperBIT can attain a depth of b = 26 mag and a total source density exceeding 40 galaxies per square arcminute. Even with the application of lensing-analysis catalog selections, we find b-band source densities between 25 and 30 galaxies per square arcminute with a median redshift of z = 1.1. Our analysis confirms SuperBIT’s capability for weak gravitational lensing measurements in the blue.}
 }
 
 @article{doi:10.1177/0278364918784361,
@@ -30,6 +38,7 @@ eprint = {
 @BOOK{AbsMahSep2008,
  author = "P.-A. Absil and R. Mahony and R. Sepulchre",
  title = "Optimization Algorithms on Matrix Manifolds",
+ doi = {10.1515/9781400830244},
  publisher = "Princeton University Press",
  address = "Princeton, NJ",
  year = 2008,
@@ -66,6 +75,7 @@ eprint = {
    author = {{Revels}, J. and {Lubin}, M. and {Papamarkou}, T.},
   journal = {arXiv:1607.07892 [cs.MS]},
      year = {2016},
[…]
