+++ paper.bib
@@ -75,7 +75,7 @@ eprint = {
    author = {{Revels}, J. and {Lubin}, M. and {Papamarkou}, T.},
   journal = {arXiv:1607.07892 [cs.MS]},
      year = {2016},
-     Doi = {https://doi.org/10.48550/arXiv.1607.07892},
+     Doi = {10.48550/arXiv.1607.07892},
       url = {https://arxiv.org/abs/1607.07892}
 }
 
@@ -381,13 +381,19 @@ abstract = {The tasks of the DAOPHOT program, developed to exploit the capabilit
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@misc{lin2023hybpsf,
-      title={HybPSF: Hybrid PSF reconstruction for the observed JWST NIRCam image}, 
-      author={Lin and Nie and Huanyuan and Shan and Guoliang and Li and Lei and Wang and Charling and Tao and Qifan and Cui and Yushan and Xie and Dezi and Liu and Zekang and Zhang},
-      year={2023},
-      eprint={2308.14065},
-      archivePrefix={arXiv},
-      primaryClass={astro-ph.IM}
+@article{lin2023hybpsf,
+doi = {10.3847/1538-3881/ad14f7},
+url = {https://dx.doi.org/10.3847/1538-3881/ad14f7},
+year = {2024},
+month = {jan},
+publisher = {The American Astronomical Society},
+volume = {167},
+number = {2},
+pages = {58},
+author = {Lin Nie and Huanyuan Shan and Guoliang Li and Lei Wang and Cheng Cheng and Charling Tao and Qifan Cui and Yushan Xie and Dezi Liu and Zekang Zhang},
+title = {HybPSF: Hybrid Point-spread Function Reconstruction for the Observed JWST NIRCam Image},
+journal = {The Astronomical Journal},
+abstract = {The James Webb Space Telescope (JWST) marks a new era of astronomical observation and discovery, providing unrivaled precision in various measurements, including photometry, morphology, and shear measurement. Accurate point-spread function (PSF) models are essential for many of these measurements. In this paper, we introduce HybPSF, a hybrid PSF construction method for JWST NIRCam imaging data that combines the WebbPSF software, which simulates the PSF for JWST, with observed data to produce more accurate and reliable PSF models. We apply this method to the SMACS J0723 imaging data and construct supplementary structures from residuals obtained by subtracting the WebbPSF PSF model from the data. Our results show that HybPSF significantly reduces discrepancies between the PSF model and the data compared to WebbPSF. Specifically, the PSF shape parameter ellipticity and size comparisons indicate that HybPSF improves precision by approximately 10 times for R 2 and 50% for e. This improvement has important implications for astronomical measurements using JWST NIRCam imaging data.}
 }
 
 @article{TheAstropyCollaboration_2022,
@@ -420,6 +426,7 @@ abstract = {The Astropy Project supports and fosters the development of open-sou
       year={2023},
       eprint={2308.10932},
       archivePrefix={arXiv},
+      doi ={10.48550/arXiv.2308.10932}
       primaryClass={astro-ph.GA}
 }
 
@@ -450,6 +457,7 @@ abstract = {The Astropy Project supports and fosters the development of open-sou
[…]
