+++ paper.bib
@@ -420,15 +420,14 @@ abstract = {The Astropy Project supports and fosters the development of open-sou
       primaryClass={astro-ph.GA}
 }
 
-@misc{franco2023cosmosweb,
-      title={COSMOS-Web: Intrinsically Luminous z$\gtrsim$10 Galaxy Candidates Test Early Stellar Mass Assembly}, 
-      author={Caitlin M. Casey and Hollis B. Akins and Marko Shuntov and Olivier Ilbert and Louise Paquereau and Maximilien Franco and Christopher C. Hayward and Steven L. Finkelstein and Michael Boylan-Kolchin and Brant E. Robertson and Natalie Allen and Malte Brinch and Olivia R. Cooper and Xuheng Ding and Nicole E. Drakos and Andreas L. Faisst and Seiji Fujimoto and Steven Gillman and Santosh Harish and Michaela Hirschmann and Shuowen Jin and Jeyhan S. Kartaltepe and Anton M. Koekemoer and Vasily Kokorev and Daizhong Liu and Arianna S. Long and Georgios Magdis and Claudia Maraston and Crystal L. Martin and Henry Joy McCracken and Jed McKinney and Bahram Mobasher and Jason Rhodes and R. Michael Rich and David B. Sanders and John D. Silverman and Sune Toft and Aswin P. Vijayan and John R. Weaver and Stephen M. Wilkins and Lilan Yang and Jorge A. Zavala},
-      year={2023},
-      eprint={2308.10932},
-      archivePrefix={arXiv},
-      doi ={10.48550/arXiv.2308.10932}
-      primaryClass={astro-ph.GA}
+@article{franco2023cosmos,
+  title={COSMOS-Web: Intrinsically Luminous z $$\backslash$gtrsim $10 Galaxy Candidates Test Early Stellar Mass Assembly},
+  author={Casey, Caitlin M and Akins, Hollis B and Shuntov, Marko and Ilbert, Olivier and Paquereau, Louise and Franco, Maximilien and Hayward, Christopher C and Finkelstein, Steven L and Boylan-Kolchin, Michael and Robertson, Brant E and others},
+  journal={arXiv preprint arXiv:2308.10932},
+      doi ={10.48550/arXiv.2308.10932},
+  year={2023}
 }
+
 
 @article{Stanitzki_2021,
 	doi = {10.1007/s41781-021-00053-3},
