+++ paper.bib
@@ -421,7 +421,7 @@ abstract = {The Astropy Project supports and fosters the development of open-sou
 }
 
 @article{franco2023cosmos,
-  title={COSMOS-Web: Intrinsically Luminous z $$\backslash$$ $$gtrsim$$ 10 Galaxy Candidates Test Early Stellar Mass Assembly},
+  title={COSMOS-Web: Intrinsically Luminous z > 10 Galaxy Candidates Test Early Stellar Mass Assembly},
   author={Casey, Caitlin M and Akins, Hollis B and Shuntov, Marko and Ilbert, Olivier and Paquereau, Louise and Franco, Maximilien and Hayward, Christopher C and Finkelstein, Steven L and Boylan-Kolchin, Michael and Robertson, Brant E and others},
   journal={arXiv preprint arXiv:2308.10932},
   doi ={10.48550/arXiv.2308.10932},
