+++ paper.bib
@@ -139,7 +139,6 @@ pages = {31},
 author = {Caitlin M. Casey and Jeyhan S. Kartaltepe and Nicole E. Drakos and Maximilien Franco and Santosh Harish and Louise Paquereau and Olivier Ilbert and Caitlin Rose and Isabella G. Cox and James W. Nightingale and Brant E. Robertson and John D. Silverman and Anton M. Koekemoer and Richard Massey and Henry Joy McCracken and Jason Rhodes and Hollis B. Akins and Natalie Allen and Aristeidis Amvrosiadis and Rafael C. Arango-Toro and Micaela B. Bagley and Angela Bongiorno and Peter L. Capak and Jaclyn B. Champagne and Nima Chartab and Óscar A. Chávez Ortiz and Katherine Chworowsky and Kevin C. Cooke and Olivia R. Cooper and Behnam Darvish and Xuheng Ding and Andreas L. Faisst and Steven L. Finkelstein and Seiji Fujimoto and Fabrizio Gentile and Steven Gillman and Katriona M. L. Gould and Ghassem Gozaliasl and Christopher C. Hayward and Qiuhan He and Shoubaneh Hemmati and Michaela Hirschmann and Knud Jahnke and Shuowen Jin and Ali Ahmad Khostovan and Vasily Kokorev and Erini Lambrides and Clotilde Laigle and Rebecca L. Larson and Gene C. K. Leung and Daizhong Liu and Tobias Liaudat and Arianna S. Long and Georgios Magdis and Guillaume Mahler and Vincenzo Mainieri and Sinclaire M. Manning and Claudia Maraston and Crystal L. Martin and Jacqueline E. McCleary and Jed McKinney and Conor J. R. McPartland and Bahram Mobasher and Rohan Pattnaik and Alvio Renzini and R. Michael Rich and David B. Sanders and Zahra Sattari and Diana Scognamiglio and Nick Scoville and Kartik Sheth and Marko Shuntov and Martin Sparre and Tomoko L. Suzuki and Margherita Talia and Sune Toft and Benny Trakhtenbrot and C. Megan Urry and Francesco Valentino and Brittany N. Vanderhoof and Eleni Vardoulaki and John R. Weaver and Katherine E. Whitaker and Stephen M. Wilkins and Lilan Yang and Jorge A. Zavala},
 title = {COSMOS-Web: An Overview of the JWST Cosmic Origins Survey},
 journal = {The Astrophysical Journal},
[…]
