+++ paper.bib
@@ -539,8 +539,3 @@ archivePrefix = {arXiv},
   pages={5340},
   year={2023}
 }
-@article{schoenholz2019jax,
-  title={Jax md: End-to-end differentiable, hardware accelerated, molecular dynamics in pure python},
-  author={Schoenholz, Samuel S and Cubuk, Ekin D},
-  year={2019}
-}
