+++ paper.bib
@@ -537,5 +537,6 @@ archivePrefix = {arXiv},
   volume={8},
   number={85},
   pages={5340},
-  year={2023}
+  year={2023},
+  doi={10.21105/joss.05340}
 }
