+++ paper.bib
@@ -451,6 +451,7 @@ abstract = {The Astropy Project supports and fosters the development of open-sou
       year={2023},
       eprint={2305.18518},
       archivePrefix={arXiv},
+      doi = {10.48550/arXiv.2305.18518},
       primaryClass={astro-ph.GA}
 }
 
