+++ paper.bib
@@ -420,6 +420,7 @@ abstract = {The Astropy Project supports and fosters the development of open-sou
       year={2023},
       eprint={2308.10932},
       archivePrefix={arXiv},
+      doi={10.48550/arXiv.2308.10932},
       primaryClass={astro-ph.GA}
 }
 
