+++ paper.bib
@@ -75,7 +75,7 @@ eprint = {
    author = {{Revels}, J. and {Lubin}, M. and {Papamarkou}, T.},
   journal = {arXiv:1607.07892 [cs.MS]},
      year = {2016},
-     Doi = {https://doi.org/10.48550/arXiv.1607.07892},
+     Doi = {10.48550/arXiv.1607.07892},
       url = {https://arxiv.org/abs/1607.07892}
 }
 
@@ -222,15 +222,18 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-
-
-@misc{rowe2015galsim,
-      title={GalSim: The modular galaxy image simulation toolkit}, 
-      author={Barnaby Rowe and Mike Jarvis and Rachel Mandelbaum and Gary M. Bernstein and James Bosch and Melanie Simet and Joshua E. Meyers and Tomasz Kacprzak and Reiko Nakajima and Joe Zuntz and Hironao Miyatake and Joerg P. Dietrich and Robert Armstrong and Peter Melchior and Mandeep S. S. Gill},
-      year={2015},
-      eprint={1407.7676},
-      archivePrefix={arXiv},
-      primaryClass={astro-ph.IM}
+@article{rowe2015galsim,
+title = {GalSim: The modular galaxy image simulation toolkit},
+journal = {Astronomy and Computing},
+volume = {10},
+pages = {121-150},
+year = {2015},
+issn = {2213-1337},
+doi = {10.1016/j.ascom.2015.02.002},
+url = {https://www.sciencedirect.com/science/article/pii/S221313371500013X},
+author = {B.T.P. Rowe and M. Jarvis and R. Mandelbaum and G.M. Bernstein and J. Bosch and M. Simet and J.E. Meyers and T. Kacprzak and R. Nakajima and J. Zuntz and H. Miyatake and J.P. Dietrich and R. Armstrong and P. Melchior and M.S.S. Gill},
+keywords = {Methods: data analysis, Techniques: image processing, Gravitational lensing, Cosmology: observations},
+abstract = {GalSim is a collaborative, open-source project aimed at providing an image simulation tool of enduring benefit to the astronomical community. It provides a software library for generating images of astronomical objects such as stars and galaxies in a variety of ways, efficiently handling image transformations and operations such as convolution and rendering at high precision. We describe the GalSim software and its capabilities, including necessary theoretical background. We demonstrate that the performance of GalSim meets the stringent requirements of high precision image analysis applications such as weak gravitational lensing, for current datasets and for the Stage IV dark energy surveys of the Large Synoptic Survey Telescope, ESA’s Euclid mission, and NASA’s WFIRST-AFTA mission. The GalSim project repository is public and includes the full code history, all open and closed issues, installation instructions, documentation, and wiki pages (including a Frequently Asked Questions section). The GalSim repository can be found at https://github.com/GalSim-developers/GalSim.}
 }
 
 @MISC{2013ascl.soft01001B,
@@ -416,6 +419,7 @@ abstract = {The Astropy Project supports and fosters the development of open-sou
       author={Ming-Yang Zhuang and Yue Shen},
       year={2023},
       eprint={2304.13776},
+      doi={10.48550/arXiv.2304.13776},
       archivePrefix={arXiv},
       primaryClass={astro-ph.GA}
 }
