+++ paper.bib
@@ -244,6 +244,7 @@ abstract = {GalSim is a collaborative, open-source project aimed at providing an
          year = 2013,
         month = jan,
           eid = {ascl:1301.001},
+          doi={ascl:1301.001},
         pages = {ascl:1301.001},
 archivePrefix = {ascl},
        eprint = {1301.001},
