+++ paper.bib
@@ -244,7 +244,6 @@ abstract = {GalSim is a collaborative, open-source project aimed at providing an
          year = 2013,
         month = jan,
           eid = {ascl:1301.001},
-          doi={ascl:1301.001},
         pages = {ascl:1301.001},
 archivePrefix = {ascl},
        eprint = {1301.001},
@@ -260,6 +259,7 @@ archivePrefix = {ascl},
        editor = {{Evans}, I.~N. and {Accomazzi}, A. and {Mink}, D.~J. and {Rots}, A.~H.},
        series = {Astronomical Society of the Pacific Conference Series},
        volume = {442},
+       doi={10.1051/aas:1996164},
         month = jul,
         pages = {435},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2011ASPC..442..435B},
