### 2024-03-21 #6144 @EdwardBerman
@ivastar I've updated all references to published versions and tracked down all of the DOI's except for one. I can't seem to find the one for PSFex anywhere, not on NASA ADS, google scholar, or any publishing site. I'm not sure it exists? Can you suggest to me a next step.

Beyond that, I have one more writing change I'd like to make per your editorial comments, and then I will have addressed everything. Thanks!
