### 2024-03-21 #6144 @ivastar
Doh. This is the only thing I can find for PSFex:

https://ui.adsabs.harvard.edu/abs/2013ascl.soft01001B/exportcitation
