+++ paper.bib
@@ -246,6 +246,7 @@ abstract = {GalSim is a collaborative, open-source project aimed at providing an
           eid = {ascl:1301.001},
         pages = {ascl:1301.001},
 archivePrefix = {ascl},
+       doi={10.1051/aas:1996164},
        eprint = {1301.001},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2013ascl.soft01001B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
