### 2024-03-24 #6144 @EdwardBerman
@ivastar I've finished going through your editorial comments and tracking down the correct DOI's. We've received our referee report for our AAS companion paper in the last couple days, and are making our way through the necessary revisions at this time. Let me know if you need anything else on our end, otherwise, I think all that is left is for my coauthors and I to work through the AAS review process. Thanks again for your help!
