### 2024-03-25 #6144 @ivastar
Excellent! Do you think that the code will change in response to the AAS comments? If not, we can may be move ahead with the post-review checklist:

https://github.com/openjournals/joss-reviews/issues/6144#issuecomment-1988409780
