+++ shopt.jl
@@ -55,6 +55,7 @@ fancyPrint("Handling Imports")
   using Dates
   using MultivariateStats
   using Base.Threads
+  using MultivariatePolynomials
   #using Interpolations
 end
 println("━ Start Time ", Dates.now())
@@ -86,9 +87,9 @@ fancyPrint("Processing Data for Fit")
 @time begin
   
   if mode == "chisq"
-    starCatalog, r, c, itr, u_coordinates, v_coordinates, outlier_indices, errVignets = cataloging(ARGS)
+    starCatalog, r, c, itr, u_coordinates, v_coordinates, outlier_indices, median_img, errVignets = cataloging(ARGS)
   else
-    starCatalog, r, c, itr, u_coordinates, v_coordinates, outlier_indices = cataloging(ARGS)
+    starCatalog, r, c, itr, u_coordinates, v_coordinates, outlier_indices, median_img = cataloging(ARGS)
   end
   #starCatalog = starCatalog
   #errVignets = errVignets
@@ -182,6 +183,8 @@ for i in sort(s_blacklist, rev=true)
   splice!(v_coordinates, i)
 end
 
+#prePixelGridFits = starCatalog
+#println(size(starCatalog[1]))
 failedStars = []
 
 # ---------------------------------------------------------#
@@ -308,6 +311,40 @@ if mode == "PCA"
 
     end
   end
+end
+
+if mode == "smooth"
+    println("━ Smooth Mode...")
+    @time begin 
+        pb = tqdm(1:length(starCatalog))
+        for i in pb
+            set_description(pb, "Star $i/$(length(starCatalog)) Complete")
+            global iteration = i
+            data = nanToZero(starCatalog[i])
+            try
+                if unity_sum
+                    try
+                        push!(pixelGridFits, smooth(data, lanczos)./sum(smooth(data), lanczos))
+                    catch
+                        #println("Smoothing failed")
+                        push!(pixelGridFits, data./sum(data))
+                    end
+                else
+                    try
+                        push!(pixelGridFits, smooth(data, lanczos))
+                    catch
+                        #println("Smoothing failed")
+                        push!(pixelGridFits, data)
+                    end
+                end
+            catch
+                println("Star $i failed")
+                push!(failedStars, i)
+                push!(pixelGridFits, zeros(r,c))
+                continue
+            end
+        end
+    end
 end
 
 
@@ -468,6 +505,23 @@ function compute_single_star_reconstructed_value(PolynomialMatrix, x, y, degree)
     return reconstructed_star
 end
 
+if median_constraint == true
+    function compute_single_star_reconstructed_value(PolynomialMatrix, x, y, degree, ppgf=median_img)
+        #println("Found pre Pixel Grid Fits")
+        r, c, _ = size(PolynomialMatrix)
+        reconstructed_star = zeros(r, c)
+        for i in 1:r
+            for j in 1:c
+                p = PolynomialMatrix[i, j, :]
+                z_data = median_img[i, j]
+                reconstructed_star[i, j] = objective_function(p, x, y, degree, median_z=z_data)
+            end
+        end
+        return reconstructed_star
+    end
+end
+    
+
 function compute_mse(reconstructed_matrix, true_matrix) #err_map
   return mean((reconstructed_matrix .- true_matrix) .^ 2 ) #./(err_map.^2)
 end
@@ -493,26 +547,20 @@ end
     @threads for i in 1:r
       for j in 1:c
         z_data = [star[i, j] for star in training_stars]
-        pC = polynomial_optimizer(degree, training_u_coords, training_v_coords, z_data)
-        PolynomialMatrix[i,j,:] .= pC
+        if median_constraint == true
[…]
