### 2024-06-18 #6144 @EdwardBerman
@ivastar Quick update, the AAS manuscript was resubmitted. I'll keep you updated on progress.
