### 2024-07-09 #6144 @EdwardBerman
@ivastar Some news on our end, we got the review on our updated AAS manuscript and it was very positive. We were told that the reviewer thinks highly of the manuscript and that they had only one edit left they wanted us to make in our discussion. When we submitted the AAS manuscript we left a note to the editor making sure they knew that we had a JOSS companion paper and what the JOSS DOI would be when accepted. I will let you know how things progress from here.
