### 2024-07-29 #6144 @EdwardBerman
@ivastar Our AAS manuscript was officially accepted! Shall we proceed with this companion paper?
