### 2024-07-29 #6144 @EdwardBerman
I archived the release on zenodo synced with the GitHub, it correctly identified myself and @mcclearyj as authors. You can find it linked here:

https://zenodo.org/records/13124495

With Markdown, it should look like this:

[![DOI](https://zenodo.org/badge/633569341.svg)](https://zenodo.org/doi/10.5281/zenodo.13124494)

The license is taken from the repo and I manually added the MIT License just in case, and I don't see anything about orcid IDs, so I'm assuming that means everything matches. The title is the same.

@ivastar Let me know if this all looks good! Thank you for all of your help through this process.
