### 2024-07-30 #6144 @ivastar
> @ivastar Looks great to me, should I run the editorial bot accept command, or is that something for you to do? Or is it just processing now that the bot added the recommend-accept label?
>

The editor in chief for the track, @dfm, has to do the actual publication magic. As noted above, the current version is still missing the AAS DOI.

> Also, is there a way to generate a tex file for the camera ready version (i.e. not `generate preprint`) so that I can post the JOSS paper to ArXiv in its final version with the same formatting as will appear in the journal? Or is the generate preprint command the default. I think its the latter, but I'd like to be sure.

We don't generate a tex file as far as I know, you can upload the PDF to arxiv.
