### 2024-07-30 #6144 @dfm
I don't think you can upload the PDF actually. They always seem to know that it was generated via latex. The above should generate what you need, I think!
