### 2024-07-30 #6144 @EdwardBerman
@augustfly I received your email, I sent back the tweaked manuscript with the appropriate citation. I'll update you all once I have the AAS DOI.
