### 2024-08-01 #6144 @EdwardBerman
@ivastar @dfm I just received my AAS DOI: 10.3847/1538-3881/ad6a0f

Shall I add this into the Markdown, or is this something to process on your end.
