### 2024-08-04 #6144 @dfm
There were a few missing journal references that I've fixed in https://github.com/EdwardBerman/shopt/pull/14. Take a look at that!
