### 2024-08-09 #6144 @dfm
Many thanks to @kevinmicha, @jpierel14 and @aymgal, for reviewing and to @ivastar for editing! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!
